Исправлен Onboarding.
Оставил только один viewController, в котором меняю backgroundView и label.
